+++ R/Plot_peak_annotation.R
@@ -25,7 +25,7 @@
 #'   package = "GenomicPlot")
 #' names(centerFile) <- c("summitPeak")
 #'
-#' bedimportParams <- list(
+#' bedimportParams <- setImportParams(
 #'   offset = 0, fix_width = 100, fix_point = "center", norm = FALSE,
 #'   useScore = FALSE, outRle = FALSE, useSizeFactor = FALSE, genome = "hg19"
 #' )