+++ NEWS.md
@@ -7,11 +7,14 @@ output:
 # CHANGES IN VERSION 0.99.12
 
 ## NEW FEATURES
-   - Add PCA plot when plot bam files correlations
+   - Add PCA plot when plot bam files correlations.
    - The function plot_5parts_metagene can generate profile plots for both
    5 parts (Promoter, 5'UTR, CDS, 3'UTR, TTS) and 3 parts (Promoter, Gene, TTS) metagene.
    
 ## SIGNIFICANT CHANGES
-   - The function plot_3parts_metagene is removed
+   - The function plot_3parts_metagene is removed.
+   - Add setImportParam function to profive default import paramters.
+   - Add saveRds option to import parameters to control saving of imported data, the default is FALSE.
    
 ## BUG FIXES
+   None